// x-pack/solutions/security/packages/ecs_data_quality_dashboard/impl/data_quality_panel/data_quality_details/indices_details/pattern/index_check_flyout/types.ts
/*
 * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
 * or more contributor license agreements. Licensed under the Elastic License
 * 2.0; you may not use this file except in compliance with the Elastic License
 * 2.0.
 */

import { type HttpHandler } from '@kbn/core-http-browser';
import { type HISTORY_TAB_ID, type LATEST_CHECK_TAB_ID } from '../constants';

export interface FetchHistoricalResultsOpts extends Partial<FetchHistoricalResultsQueryState> {
  indexName: string;
  httpFetch: HttpHandler;
  abortController: AbortController;
}

export type UseHistoricalResultsFetchOpts = Omit<FetchHistoricalResultsOpts, 'httpFetch'>;

export type UseHistoricalResultsFetch = (opts: UseHistoricalResultsFetchOpts) => Promise<void>;

export interface FetchHistoricalResultsQueryState {
  from: number;
  size: number;
  startDate: string;
  endDate: string;
  outcome?: 'pass' | 'fail';
}

export type IndexCheckFlyoutTabId = typeof HISTORY_TAB_ID | typeof LATEST_CHECK_TAB_ID;